Add bstr dependency
This crate is intended to be consumed by `artichoke-backend` to
implement the `digest` stdlib module. In this context, `bstr` is an
acceptable dependency to take since `artichoke-backend` already depends
on it, and likely will continue to depend on it.

`bstr` also gets this crate access to high performance byte manipulation
routines which should improve encode and decode performance.