Improve error handling in boba::decode
`boba::decode` validates that the bytestring only contains ASCII bytes
from the encoding alphabet and uses this to justify unwrapping when
converting bytes to indexes in the `VOWELS` and `CONSONTANTS` byte
arrays.

This is flawed because a bytestring can pass the initial validation step
if it uses vowels where a consonant is expected or consonants where a
vowel is expected.

This commit introduces two new error variants to `DecodeError` for cases
where the decode step encounters a byte in the other partition of the
encoding alphabet.

Two new tests are added.