Scope encoded length to block for constructing the decode buffer
`boba::decode` eagerly allocates a `Vec` for output with
`Vec::with_capacity` using the length of the encoded input. To avoid
calling `encoded.len()` twice in the capacity computation, it is stored
in a `len` binding.

This binding is otherwise unused. Scoping the binding to a block for its
single use makes the intent clearer and prevents the variable from
accidentaly being used later.