Minimize codebloat by moving encoder and decoder out of generic function
Move encode implementation to an internal function that takes a `&[u8]`.
The public `boba::encode` is generic on `T: AsRef<[u8]>`. This public
function now only calls `as_ref` and immediately delegates, which will
minimize the amount of duplicate codegen if `encode` is called with
`Vec<u8>` and `&str`, etc.

The same change is made for `boba::decode`.

Internal encoding and decoding arithmetic routines are moved to the new
`encode` and `decode` modules.