Optimize `boba::decode`
Replace casting in and out of `usize` and `u8` with lookup functions
from alphabet to `Option<u8>`. This speeds up `decode` performance by
40%.

I tried a similar approach with the inverse lookup functions in
`encode`, but it ended up regressing performance. I also tried turning
the `VOWELS` and `CONSONANTS` byte arrays into `&str` and using string
slicing + `Vec::push_str`, which also was a perf regression, though not
as terrible.

This change moves the constants to the impls that use them.