Remove `bstr` and reimplement bits that are used in `decode`
The only remaining use of `bstr` in `boba` was a single call to
`ByteSlice::find_not_byteset` in `boba::decode::inner`. After looking at
the `bstr` source, for an alphabet the size of bubblebabble's encoding,
it computed a `[u8; 256]` lookup table and checks for membership with
`O(1)` array access.

Because we know the alphabet ahead of time, we can pre-compute the table
and implement this with `Iterator::find`.

This results in a modest 3% speedup in decode performance, mostly because
the alphabet lookup table is a pre-computed const.