Remove `alloc` feature
`boba` has a required dependency on the `alloc` crate, but gates the
availability of the `alloc` crate with the `alloc` feature.

`boba` does not build with no default features as a result. The docs
lead me to believe this was intentional, but with new CI steps in #159,
all Artichoke crates must build and pass a docs build with no default
features.

This will require a semver major bump.